Update and rename InternetConectivity.java to InternetConnectivity.java
